requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
mysqlclient==2.2.1
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.8.3
Werkzeug==3.0.1
gunicorn==21.2.0
supabase==2.10.0
requests==2.31.0
Pillow==10.1.0
redis==5.0.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
flasgger==0.9.7.1
